BREAKING CHANGE: Replace homemade logging function
This removes the 'message' function along with all of the 'verbosity'
arguments in all of the functions.

The built-in 'logging' module is used instead of print statements and
the the 'JournalHandler' supplied by the systemd module is used to
pass messages to the systemd journal.

This doesn't change anything outwardly apart from what stream is used.
Everything used to go to STDOUT, whereas now all messages go to STDERR.

The only change is if you use any of the functions directly.