Display messages by default
This used to be set to hide them because it was usually run as a service
and before bf4f91e4124a41b6a672f4fd0d5654d2d836a0e4 the messages were
sent to the journal regardless. Since then though, unless you enabled
the message printing, nothing would be sent to the journal. Doing this
however, meant that you would get duplicate logs.

This fixes the above issue by disabling the STDOUT and STDERR forwarding
to the systemd journal.