MAJOR: allow not saving output files
For all of the output generators, the 'output_file' argument now defaults
to being an empty string, this in turn disables the saved output.

This also changes the allowable strings for disabling the output
to include an empty one so that you can leave a line blank
in the configuration file to disable a generator.